# www/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Terminal PDV e ERP - API corporativa",
        "description": "Documentação OpenApi para integração do ERP Curitiba Serviços. Endpoints cobrem RH, Vendas, Inventário e Financeiro.",
        "contact": {
            "email": "suporte@curitibaservicos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor Primário"
        }
    ],
    "paths": {
...
        "/api/hr/employees": {
            "get": {
                "tags": [
                    "Recursos Humanos"
                ],
                "summary": "Listar Funcionários",
                "description": "Retorna a lista de colaboradores incluindo dados de folha e status de acesso.",
                "operationId": "518d03cd473804baa2d8662f4508afc9",
                "responses": {
                    "200": {
                        "description": "Operação bem-sucedida"
                    }
                }
            },
            "post": {
                "tags": [
                    "Recursos Humanos"
                ],
                "summary": "Registrar Funcionário",
                "description": "Cadastra um novo colaborador no sistema",
                "operationId": "3a7562bf719063d0738f8861bca577b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Maria Souza"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "12345678909"
                                    },
                                    "role_description": {
                                        "type": "string",
                                        "example": "Caixa Master"
                                    },
                                    "base_salary_cents": {
                                        "type": "integer",
                                        "example": 350000
                                    },
                                    "pin": {
                                        "type": "string",
                                        "example": "4455"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado"
                    }
                }
            }
        },
...
    },
...
}